sdks/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "micelio"
version = "0.1.0"
description = "Python SDK for the Agent Internet Protocol (AIP)"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "cryptography>=41.0",
    "python-ulid>=2.0",
]

[project.optional-dependencies]
discovery = ["zeroconf>=0.80"]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["micelio*"]
